services/backend/code/requirements.txt
Flask
psycopg2-binary
flask-cors
flask-sqlalchemy
flask-login
flask-jwt-extended